// cinematch-frontend/src/models/history/movie-action-type.ts
export enum MovieActionType {
    LIKE = "LIKE",
    UNLIKE = "UNLIKE",
    WATCHED = "WATCHED",
    UNWATCHED = "UNWATCHED",
    WATCHLIST_ADD = "WATCHLIST_ADD",
    WATCHLIST_REMOVE = "WATCHLIST_REMOVE",
    REVIEW_ADD = "REVIEW_ADD",
    REVIEW_EDIT = "REVIEW_EDIT",
    REVIEW_REMOVE = "REVIEW_REMOVE",
}